requirements.txt
Flask==3.0.0
pandas==1.5.0
Pillow==9.3.0
python-docx==1.1.0
requests==2.28.1